+++ Faker/Provider/Base.php
@@ -1,0 +1,17 @@
+<?php
+
+
+namespace Kaliop\eZLoremIpsumBundle\Faker\Provider;
+
+use Faker\Generator;
+
+abstract class Base
+{
+    /** @var Generator $generator */
+    protected $generator;
+
+    public function __construct(Generator $generator)
+    {
+        $this->generator = $generator;
+    }
+}